requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
accelerate=0.28.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=py312h06a4308_0
annotated-types=0.6.0=pypi_0
arrow-cpp=14.0.2=h374c478_1
attrs=23.1.0=py312h06a4308_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.10.55=h721c034_0
bitarray=2.9.2=pypi_0
bitsandbytes=0.43.0=pypi_0
blas=1.0=mkl
blinker=1.6.2=py312h06a4308_0
blis=0.7.11=pypi_0
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.7=py312ha883a20_0
brotli-python=1.0.9=py312h6a678d5_7
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py312h06a4308_0
cloudpathlib=0.16.0=pypi_0
colbert-ai=0.2.19=pypi_0
confection=0.1.4=pypi_0
cuda-cudart=12.4.99=0
cymem=2.0.8=pypi_0
datasets=2.18.0=pypi_0
dill=0.3.8=pypi_0
en-core-web-sm=3.7.1=pypi_0
expat=2.5.0=h6a678d5_0
faiss-gpu=1.8.0=py3.12_h4c7d538_0_cuda12.1.1
filelock=3.9.0=pypi_0
flask=3.0.2=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
gflags=2.2.2=h6a678d5_1
git-python=1.0.3=pypi_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.37=py312h06a4308_0
glog=0.5.0=h6a678d5_1
grpc-cpp=1.48.2=he1ff14a_1
huggingface-hub=0.21.4=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=py312h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.19.2=py312h06a4308_0
jsonschema-specifications=2023.7.1=py312h06a4308_0
krb5=1.20.1=h143b758_1
langcodes=3.3.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=12.4.2.65=0
libcurl=8.5.0=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libfaiss=1.8.0=h046e95b_0_cuda12.1.1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=2.2.0=py312h06a4308_1
markupsafe=2.1.3=py312h5eee18b_0
mdurl=0.1.0=py312h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.10=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
numexpr=2.8.7=py312hf827012_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.19.3=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=23.2=py312h06a4308_0
pandas=2.2.1=py312h526ad5a_0
pillow=10.2.0=py312h5eee18b_0
pip=23.3.1=py312h06a4308_0
preshed=3.0.9=pypi_0
protobuf=3.20.3=py312h6a678d5_0
psutil=5.9.8=pypi_0
pyarrow=14.0.2=py312hb107042_0
pyarrow-hotfix=0.6=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pydeck=0.8.0=py312h06a4308_2
pygments=2.15.1=py312h06a4308_1
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=h996f2a0_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.1=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py312h06a4308_0
pyyaml=6.0.1=py312h5eee18b_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
referencing=0.30.2=py312h06a4308_0
regex=2023.12.25=pypi_0
requests=2.31.0=py312h06a4308_1
rich=13.3.5=py312h06a4308_1
rpds-py=0.10.6=py312hb02cf49_0
s2n=1.3.27=hdbd6064_0
safetensors=0.4.2=pypi_0
scipy=1.12.0=pypi_0
setuptools=68.2.2=py312h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.4.0=pypi_0
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.10=h6a678d5_1
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
streamlit=1.32.0=py312h06a4308_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.2=py312h06a4308_1
thinc=8.2.3=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py312h06a4308_0
torch=2.2.1+cu118=pypi_0
torchaudio=2.2.1+cu118=pypi_0
torchvision=0.17.1+cu118=pypi_0
tornado=6.3.3=py312h5eee18b_0
tqdm=4.66.2=pypi_0
transformers=4.39.0=pypi_0
typer=0.9.0=pypi_0
typing_extensions=4.9.0=py312h06a4308_1
tzdata=2024a=h04d1e81_0
ujson=5.9.0=pypi_0
urllib3=2.1.0=py312h06a4308_1
utf8proc=2.6.1=h5eee18b_1
wasabi=1.1.2=pypi_0
watchdog=2.1.6=py312h06a4308_0
weasel=0.3.4=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py312h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
